services/agents_service/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-filter>=24.2
psycopg[binary]>=3.2
celery>=5.4
redis>=5.0
django-prometheus>=2.3
